Add solution for MissingNumbers
We are taking advantage the difference between elements is guaranteed to be less than 100 (i.e. lots of duplicates).
We can sort the keys on insertion, then print the results in ascending order simply by iterating over the remaining keys.
If the range of keys was much bigger, it would be more efficient to use a HashMap and sort the keys at the end, just before printing.